+++ src/BlazorLazyLoad/LazyLoadComponentPlaceHolder.cs
@@ -26,6 +26,8 @@ namespace BlazorLazyLoad
 			_parms=parms.Where(x => string.Equals(x.Key,"_Assembly",StringComparison.OrdinalIgnoreCase)==string.Equals(x.Key,"_Type",StringComparison.OrdinalIgnoreCase)).ToArray();
 
 			_innerComponentType=await EnsureComponentAsync(asm,type);
+			if (_innerComponentType==null)
+				throw new ArgumentException($"Assembly {asm} can't be resolved.");
 
 			StateHasChanged();
 		}
@@ -41,9 +43,29 @@ namespace BlazorLazyLoad
 
 		async Task<Type> EnsureComponentAsync(string assemblyName,string type)
 		{
-			IEnumerable<Assembly> newAssemblies = (await AssemblyDependencyResolver.ResolveAsync(assemblyName));
-			LoadServices(newAssemblies);
-			return newAssemblies.First().GetType(type,true);
+			//We need to inject new assembly to the router because it resolves which page to display.
+			IRouterEnvelope router = JSInteropMethods.Router;
+			IEnumerable<Assembly> additionalAssemblies = router.AdditionalAssemblies??Enumerable.Empty<Assembly>();
+
+			//Don't inject the assembly multiple times.
+			Assembly asm = additionalAssemblies.FirstOrDefault(x => string.Equals(x.GetName().Name,assemblyName,StringComparison.OrdinalIgnoreCase));
+			if (asm==default)
+			{
+				//Load assembly including its dependencies
+				IEnumerable<Assembly> newAssemblies = await AssemblyDependencyResolver.ResolveAsync(assemblyName);
+				if (!newAssemblies.Any())
+					return null;
+
+				LoadServices(newAssemblies);
+
+				//Inject the assembly to the router.
+				ParameterView pv = ParameterView.FromDictionary(new Dictionary<string,object>() { { nameof(IRouterEnvelope.AdditionalAssemblies),additionalAssemblies.Concat(newAssemblies).ToArray() } });
+				await router.SetParametersAsync(pv);
+
+				asm=newAssemblies.First();
+			}
+
+			return asm.GetType(type,true);
 		}
 
 		protected virtual void LoadServices(IEnumerable<Assembly> newAssemblies)